[genesis] Init genesis in db, not state
